(PUP-10252) Remove Puppet::Util::SSL
The subject_from_dn method is leftover from webrick/passenger and the
handle_connection_error method is unused now that puppet uses
Puppet::SSL::Verifier exclusively.

Move the remaining cn_from_subject to Puppet::SSL::Base which is the only caller.
